Re-organize stack and flamegraph related code
Move the more general code to the flamegraphs package and
leave only the conversion from jfr files to stacks in the
jfr package.
